Store a Fedora version of hardware:razer.repo
This uses the `Fedora_$releasever` variable so the user is always
pointed to the correct release repository for their Fedora version.

Works for: Fedora 39, 40, 41

Does not work: Rawhide
    (The URL request is lowercase)

However, the command to add repo changed in Fedora 41, so we will need
to keep the two for now until Fedora 40 is obsolete.

Closes #1981